Single-source "version", bump to 2.0.0 production
1.  Apart from the textual README.md file, which still must be manually
adjusted, all other references to "version" come from the __version__
variable set in geeknote/__init__.py

2.  Bump version to 2.0.0, and adopt the triplet version scheme.

3.  Alter the package details in setup.py, so that I (Jeff Kowalski)
am now the author, and the website url points to my github repo.  As
well the Development status is now 5 (production/stable).  Please note
that full credit to the original authors is given in the README.md.  I'm
not meaning to claim credit for authorship, but simply give users some
clarity on maintenance.