Prompt for developer token when login fails
This is a reaction to jeffkowalski/geeknote #89. Evernote has revoked
geeknote's API credentials, and for now at least it's necessary for each
user individually to use a personal developer token. It's only necessary
to provide it once, via environment, but that's still somewhat buried.

This update detects when the OAuth temporary token api call fails, and
prompts for a developer token. Once user has entered one, it's retained
in the client settings database and doesn't need to be provided again.