Updated readme on obtaining developer tokens
As seen on #89, developer tokens are harder to get that one expects, and they
must be converted to user tokens.

As normal login via oauth is currently disabled, this is the only way to login
and it requires a complex process. This commit documents the process referring
to the FAQ in a step-by-step manner.