Correct storage of default note extensions (fix issue #44)
Formerly, default note extensions were stored as a string, which
presented problems due to possible embedded spaces between elements.
For example, ".markdown, .org" would have been interpreted as
".markdown" and " .org".  Now, the extensions are stored as a list of
strings ['.markdown', '.org'], even though they're still set by the
user as a comma-delimited string.

NOTE BEFORE UPGRADING:
This change requires you re-set the default note extensions if you
have customized them, or they'll be reset to the defaults
1) It's best to first record your current settings with
  geeknote settings --note_ext
2) Now upgrade geeknote
3) Set the extensions again after upgrading, like
  geeknote settings --note_ext ".markdown, .org"

This fixes #44 - Edit a note with raw mode, opened two empty files
The extra file showed up when editing in raw mode due to the extra
space after the comma in the string version of the default note
extensions.  This made the editing command appear to have two files:
the actual temp file and another called ".org". The normal markdown
version did not have a space and so it was appended to the
temporary file name correctly.