Fix issue #121 - notes with resources but no content weren't created
When creating a note, but not explicitly providing content, geeknote
assumes the editor should be used to collect user input.  It then
creates a note only when the content is changed.  If the content
remains empty, no note would have been created.

Now, we create the new note anyway even when the empty content is unchanged.
Perhaps a better fix would be to create the note only if a resource
were provided in that case.  I'm not sure that's obvious behavior though.