workaround total note count evernote bug
Evernote api can return a total note count that is 16 greater than the
actual length of the list of available notes.  So, rather than taking
total notes as a target in the fetching loop, we also look to see if
the retrieved notes length is zero, at which point we presume there
will be no further notes.